backend/requirements.txt
flask
gunicorn
flask-cors
requests
cassandra-driver
